+++ app/config/index.js
@@ -4,14 +4,16 @@ const Joi = require('joi')
 const schema = Joi.object({
   serviceName: Joi.string().default('Request Editor'),
   port: Joi.number().default(3001),
-  env: Joi.string().valid('development', 'test', 'production').default('development')
+  env: Joi.string().valid('development', 'test', 'production').default('development'),
+  robotSheetName: Joi.string().required()
 })
 
 // Build config
 const config = {
   serviceName: process.env.SERVICE_NAME,
   port: process.env.PORT,
-  env: process.env.NODE_ENV
+  env: process.env.NODE_ENV,
+  robotSheetName: process.env.ROBOT_SHEET_NAME
 }
 
 // Validate config
